INS-113: Merge types.go into record.go in record module.
